+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Change Log
 All notable changes to this project will be documented in this file.
-This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/xsolla/xsolla-sdk-php/compare/v2.0.0-BETA1...master)
 ### Added
@@ -13,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Added optional argument `$sandboxMode` to `XsollaClient::createCommonPaymentUIToken`
 * [BC BREAK] `XsollaClient::ListSubscriptionPayments` returns subscription payments for all users; `user_id` parameter removed. Added `ListUserSubscriptionPayments` instead `ListSubscriptionPayments` with same parameters.
+* [BC BREAK] Changed `plan_id` type in Subscriptions API from `string` to `int`
 
 ## [v2.0.0-BETA1](https://github.com/xsolla/xsolla-sdk-php/compare/v1.1.1...v2.0.0-BETA1)
 ### Added